flask&folder updates post sprint meeting
