--- src/visualizations.py
@@ -1,1 +1,0 @@
-##visualisations.py will be used to create visualizations for the flask app program.
